+++ examples/basic/mousehover1.py
@@ -14,10 +14,12 @@ def func(evt):                       ### called every time mouse moves!
           f"Height: {precision(arr[pid],3)}\n" \
           f"Ground speed: {precision(evt.speed3d*100,2)}"
     msg.text(txt)                    # update text message
+
     arw = Arrow(pt - evt.delta3d, pt, s=0.001, c='orange5')
-    fp = msh.flagpole(txt, point=pt, offset=(0.4,0.6),
-                      s=0.04, c='k', font="VictorMono")
-    fp.follow_camera()               # make it always face the camera
+    fp = msh.flagpole(
+        txt, point=pt, offset=(0.4,0.6),
+        s=0.04, c='k', font="VictorMono",
+    ).follow_camera()  # make it always face the camera
     if len(plt.actors) > 3:
         plt.pop()                    # remove the old flagpole
     plt.add(arw, fp)                 # add Arrow and the new flagpole
@@ -27,10 +29,10 @@ hil = ParametricShape('RandomHills').cmap('terrain').add_scalarbar()
 arr = hil.pointdata["Scalars"]       # numpy array with heights
 
 plt = Plotter(axes=1, bg2='lightblue')
+
 plt.add_callback('mouse move', func) # add the callback function
-plt.add_callback('keyboard',
-                 lambda evt: plt.remove(plt.actors[3:]).render(),
-)
+plt.add_callback('keyboard', lambda evt: plt.remove(plt.actors[3:]).render())
+
 plt.show(hil, msg, __doc__, viewup='z')
 plt.close()