+++ examples/pyplot/goniometer.py
@@ -9,10 +9,13 @@ mesh = Cone().c("steelblue").rotate_y(90).pos(1, 2, 3)
 # add a flagpole-style comment
 a, v = precision(mesh.area(), 4), precision(mesh.volume(), 4)
 fp = mesh.flagpole(
-    "S = πr^2 +πr√(h^2 +r^2 )\n  = " + a + "~μm^2 \nV = πr^2 ·h/3\n  = " + v + "~μm^3",
+    "S = πr^2 +πr√(h^2 +r^2 )\n  = " + a 
+    + "~μm^2 \nV = πr^2 ·h/3\n  = " + v 
+    + "~μm^3",
     s=0.1,
 )
-fp.color("r3").scale(0.7).follow_camera()
+
+fp.color("r3").scale(0.7)
 
 # measure the angle formed by 3 points
 gon = Goniometer(
@@ -32,4 +35,4 @@ rul = Ruler(
 # make 3d rulers along the bounding box (similar to set axes=7)
 ax3 = RulerAxes(mesh, units="μm")
 
-show(mesh, fp, gon, rul, ax3, __doc__, bg2="lb", viewup="z").close()
+show(mesh, fp.follow_camera(), gon, rul, ax3, __doc__, bg2="lb", viewup="z").close()